+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased][unreleased]
 
+## [0.5.0] - 2023-03-29
 ### Added
 - Added rule for checking required parameters in method calls
 - Added forbidden constructor parameters types rule
@@ -42,7 +43,8 @@
 - Rule for checking trait context 
 - Phpstorm meta dynamic return types
 
-[unreleased]: https://github.com/efabrica-team/phpstan-rules/compare/0.4.2...HEAD
+[unreleased]: https://github.com/efabrica-team/phpstan-rules/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/efabrica-team/phpstan-rules/compare/0.4.2...0.5.0
 [0.4.2]: https://github.com/efabrica-team/phpstan-rules/compare/0.4.1...0.4.2
 [0.4.1]: https://github.com/efabrica-team/phpstan-rules/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/efabrica-team/phpstan-rules/compare/0.3.0...0.4.0